Generalize computing of paths from user strings
This takes the previous default_file_name_for_glyph_name
function out of the glyph module and moves it to a new
`util` module; it then adapts it to also work for layer names.
